Moved user handlers into separate files for handlers and users for clarity
